==> fig4/00.organize_data_random.ipynb <==
import sys
sys.path.insert(0, '/cndd/fangming/CEMBA/snmcseq_dev')
from scipy import sparse
# import scipy
from scipy import io

from __init__ import *
from __init__jupyterlab import *
import snmcseq_utils

# obsolete

f = '/cndd/Public_Datasets/CEMBA/BICCN_minibrain_data/data_freeze/snmc_and_snatac/study/atac/yang_analysis_20190912/MOp.GABAergic.MGE.cluster.meta.txt'
df = pd.read_csv(f, sep='\t').rename(columns={'x.sp@sample': 'sample', 
                                              'x.sp@cluster': '_cluster', 
                                             })
df['cell'] = df['sample'] + '_' + df['barcode']

f = '/cndd/Public_Datasets/CEMBA/BICCN_minibrain_data/data_freeze/snmc_and_snatac/study/atac/yang_analysis_20190912/MOp.GABAergic.MGE.cluster.anno.txt'
df_anno = pd.read_csv(f, sep='\t').rename(columns={'cluster': '_cluster', 
                                                   'anno': 'sub_cluster', 
                                                })
df = pd.merge(df, df_anno, on='_cluster')
df = df.set_index('cell')

print(df.shape)
df.head()


print(len(df['sub_cluster'].unique()))

f = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_formatted/snatac_gene_metadata.tsv'
meta = pd.read_csv(f, sep='\t', index_col=0)
meta['sub_cluster'] = meta['cluster']
meta.update(df[['sub_cluster']])

print(len(meta['cluster'].unique()))
print(len(meta['sub_cluster'].unique()))
print(meta.shape)
meta.head()

meta.update(df[['sub_cluster']])

df[['sub_cluster']].head()

# atac-2kb 

f = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/count.mm'
mat = io.mmread(f)
print(mat.shape)
mat

f = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/count.gene'
genes = pd.read_csv(f)['x'].values
genes = np.array([gene.split('.')[0] for gene in genes])
print(genes.shape)
genes

f = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/count.barcode'
cells = pd.read_csv(f)['x'].values
print(cells.shape)
cells

f = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/count.sample'
samples = pd.read_csv(f)['x'].values
samples = np.array([sample.split('/')[-1].split('.')[0] for sample in samples])
print(samples.shape)
samples

barcodes = np.array(['snatac_gene_'+sample+'_'+cell for sample, cell in zip(samples, cells)])
print(barcodes.shape)
barcodes

gc_mat = GC_matrix(genes, barcodes, mat.T)
gc_mat

f_gene = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/raw.gene'
f_cell = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/raw.cell'
f_mat = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/raw.npz'
snmcseq_utils.save_gc_matrix(gc_mat, f_gene, f_cell, f_mat)

##  get neuronal cells
- and save file  
- and get hvfeatures

f_gene = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/count.gene'
f_cell = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/count.cell'
f_mat = '/cndd/fangming/CEMBA/scripts/from_wayne/run_process_snapATAC/run/snap_files_analyzed/count.npz'
gc_mat = snmcseq_utils.load_gc_matrix(f_gene, f_cell, f_mat)

f = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_neurons_v2/snatac_gene_metadata.tsv'
meta = pd.read_csv(f, sep="\t")
print(meta.shape)
meta.head()

gc_mat.cell.shape

selected_cells_idx = snmcseq_utils.get_index_from_array(gc_mat.cell, meta['cell'].values)
assert np.all(selected_cells_idx != -1)


gc_mat_update = GC_matrix(gc_mat.gene, 
                          gc_mat.cell[selected_cells_idx], 
                          gc_mat.data.tocsc()[:, selected_cells_idx]
                         )

data_dir = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_neurons_v2'
f_gene = os.path.join(data_dir, 'snatac_gene_raw.gene')
f_cell = os.path.join(data_dir, 'snatac_gene_raw.cell')
f_mat = os.path.join(data_dir, 'snatac_gene_raw.npz')
snmcseq_utils.save_gc_matrix(gc_mat_update, f_gene, f_cell, f_mat)
