==> fig4/01-Copy1-mcg.prep_data_format.ipynb <==
# Organize freezed data
Aug 19th, 2019

Sep 13th, 2019 (data freeze)

Fangming Xie

- reformat raw data (from the dropbox) into standardized python readable format
- use ensembl id to refer genes
- preserve content as much as possible, including naming conventions of table header and index

import sys
sys.path.insert(0, '/cndd/fangming/CEMBA/snmcseq_dev')
import importlib

from __init__ import *
from __init__jupyterlab import *
import snmcseq_utils
importlib.reload(snmcseq_utils)
import CEMBA_clst_utils
from scipy.io import mmread
from scipy import sparse
import time


SRC_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze'
DST_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_formatted'

f = PATH_GENEBODY_ANNOTATION 
df_genes = pd.read_csv(f, sep="\t")
df_genes['gene_id'] = df_genes['gene_id'].apply(lambda x: x.split('.')[0])
df_genes = df_genes.set_index('gene_id')
df_genes_v2 = df_genes.reset_index().groupby('gene_name').first()
print(df_genes.shape)
df_genes.head()
print(df_genes_v2.shape)
df_genes_v2.head()

### mC

- QC'ed cells
- remove version on ensembl id

import xarray as xr

dataset_name = 'snmcseq_gene'

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data_mc = os.path.join(DST_DIR, dataset_name+'_mCG_counts.npz') 
fout_data_c = os.path.join(DST_DIR, dataset_name+'_CG_counts.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_counts.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_counts.cell') 

print(fout_meta)
print(fout_data_mc)
print(fout_data_c)

# metadata after QC and clustering
# meta = pd.read_msgpack(os.path.join(SRC_DIR, 'snmc_and_snatac/study/mc/Metadata.after_QC.msg'))
# meta2 = pd.read_msgpack(os.path.join(SRC_DIR, 'snmc_and_snatac/study/mc/mop_mc_clustering/MOp.methylome_only.analysis_results.msg')).drop(columns=['cell'])
# print(meta.shape)
# print(meta2.shape)
# print(meta.columns.values.shape)
# print(np.intersect1d(meta.columns.values, meta2.columns.values).shape)
# meta2 = meta2.drop(columns=meta2.filter(regex='^ENSMUSG', axis=1).columns).drop(columns='AllcPath')

meta3 = pd.read_csv(os.path.join(SRC_DIR, 'snmc_and_snatac/study/mc/MOp.snmC-seq.AnalysisResult.csv.gz')).rename(columns={'index': 'cell'})
print(meta3.shape)
meta3.head()

# meta3.to_csv(fout_meta, sep="\t", header=True, index=False)

reps = ['2C-180409.mcds', 
        '2C-180410.mcds',
        '3C-171206.mcds',
        '3C-171207.mcds',
        '4B-171212.mcds',
        '4B-171213.mcds',
        '4B-180104.mcds',
        '5D-180605.mcds',
        '5D-180612.mcds',
       ]

c_tables = []
mc_tables = []
for rep in reps:
    mcds = xr.open_dataset(os.path.join(SRC_DIR, 'snmc_and_snatac/dataset/mc/{}'.format(rep)))
    c_table = mcds['gene_da'].sel(mc_type='CGN', count_type='cov').squeeze().to_pandas()
    mc_table = mcds['gene_da'].sel(mc_type='CGN', count_type='mc').squeeze().to_pandas()
    c_tables.append(c_table)
    mc_tables.append(mc_table)
    
c_tables = pd.concat(c_tables, axis=0)
mc_tables = pd.concat(mc_tables, axis=0)
assert np.all(c_tables.columns.values == mc_tables.columns.values)
assert np.all(c_tables.index.values == mc_tables.index.values)
print(c_tables.shape, mc_tables.shape)
    
gene = c_tables.columns.values
cell = c_tables.index.values 
c_table = c_tables.T
c_table.index = gene

mc_table = mc_tables.T
mc_table.index = gene

# remove version on ensembl id
c_table = sparse.coo_matrix(c_table.values)
mc_table = sparse.coo_matrix(mc_table.values)

cell = c_tables.index.values
cell

# save mc and c tables

data = {'mc': mc_table, 
        'c': c_table, 
       }
gc_mat = GC_matrix(gene, cell, data) 

# snmcseq_utils.save_gc_matrix_methylation(gc_mat, 
#                                          fout_data_gene, fout_data_cell, 
#                                          fout_data_mc, fout_data_c)

### mC

- QC'ed cells
- remove version on ensembl id

SRC_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_formatted'
DST_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_qc_cells'

dataset_name = 'snmcseq_gene'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data_mc = os.path.join(SRC_DIR, dataset_name+'_mCG_raw.npz') 
f_data_c = os.path.join(SRC_DIR, dataset_name+'_CG_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(f_data_mc)
print(f_data_c)
print(f_data_gene)
print(f_data_cell)

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data_mc = os.path.join(DST_DIR, dataset_name+'_mCG_raw.npz') 
fout_data_c = os.path.join(DST_DIR, dataset_name+'_CG_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 


print(fout_meta)
print(fout_data_mc)
print(fout_data_c)
print(fout_data_gene)
print(fout_data_cell)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
# # meta = meta[meta['MajorCluster']!='Outlier']
# print(meta.groupby('MajorCluster').size())
print(meta.shape)
meta.head()

# gc_mat = snmcseq_utils.load_gc_matrix_methylation(f_data_gene, f_data_cell, f_data_mc, f_data_c)
print(gc_mat.data['mc'].shape, gc_mat.data['c'].shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([cell for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells
assert np.all(selected_cells_id != -1)

data_update = {'mc': gc_mat.data['mc'].tocsc()[:, selected_cells_id], 
               'c': gc_mat.data['c'].tocsc()[:, selected_cells_id], 
              }

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, data_update)
print(gc_mat_update.data['mc'].shape, 
      gc_mat_update.data['c'].shape, 
      gc_mat_update.gene.shape, gc_mat_update.cell.shape)

### mC

- QC'ed cells
- remove version on ensembl id

SRC_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_qc_cells'
DST_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_neurons'

dataset_name = 'snmcseq_gene'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data_mc = os.path.join(SRC_DIR, dataset_name+'_mCG_raw.npz') 
f_data_c = os.path.join(SRC_DIR, dataset_name+'_CG_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(f_data_mc)
print(f_data_c)
print(f_data_gene)
print(f_data_cell)

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data_mc = os.path.join(DST_DIR, dataset_name+'_mCG_raw.npz') 
fout_data_c = os.path.join(DST_DIR, dataset_name+'_CG_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 


print(fout_meta)
print(fout_data_mc)
print(fout_data_c)
print(fout_data_gene)
print(fout_data_cell)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.shape)
meta.head()

# gc_mat = snmcseq_utils.load_gc_matrix_methylation(f_data_gene, f_data_cell, f_data_mc, f_data_c)
print(gc_mat.data['mc'].shape, gc_mat.data['c'].shape, gc_mat.gene.shape, gc_mat.cell.shape)

# meta[meta]
print(meta.groupby('MajorCluster').size())
print(meta.shape)
meta = meta.loc[meta['MajorCluster'].str.contains(r'^L|^MGE|^CGE')] # removed non-neuronal and outlier
print(meta.shape)
meta['MajorCluster'].unique()


selected_cells = meta.index.values
selected_cells_id = snmcseq_utils.get_index_from_array(['snmcseq_gene_' + i for i in gc_mat.cell], 
                                                       selected_cells)
assert np.all(selected_cells_id != -1)

data_update = {'mc': gc_mat.data['mc'].tocsc()[:, selected_cells_id], 
               'c': gc_mat.data['c'].tocsc()[:, selected_cells_id], 
              }

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, data_update)
print(gc_mat_update.data['mc'].shape, 
      gc_mat_update.data['c'].shape, 
      gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# save mc and c tables

# snmcseq_utils.save_gc_matrix_methylation(gc_mat_update, 
#                                          fout_data_gene, fout_data_cell, 
#                                          fout_data_mc, fout_data_c)

# gc_mat_update.gene[-3:] #.shape
# gc_mat_update.cell.shape
# gc_mat_update.cell[-3:]

# Settings

SRC_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_neurons'
DST_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_neurons'
sys.path.insert(0, SRC_DIR)
from __init__datasets import *

f_meta_format = '{0}/{1}_metadata.tsv'
f_data_format = '{0}/{1}_{2}raw.{3}'
f_hvftr_format = '{0}/{1}_hvfeatures.{2}'
f_ftr_format = '{0}/{1}_features.{2}'


def get_size_in_GB(obj):
    """"""
    GB = 1024**3
    return sys.getsizeof(obj)/GB

mods_selected = [
    'snmcseq_gene',
#     'snatac_gene',
    
#     'smarter_nuclei',
#     'smarter_cells',
#     '10x_nuclei_v3_macosko',
#     '10x_cells_v3',
#     '10x_nuclei_v3',
#     '10x_cells_v2',
    
#     '10x_nuclei_v2',
#     'smarter-cells-v1',
]

# # gene name as index
# gene_annot_file = PATH_GENEBODY_ANNOTATION
# gene_annot = pd.read_csv(gene_annot_file, sep="\t")
# gene_annot_v2 = gene_annot.groupby('gene_name').first()
# print(gene_annot_v2.shape)
# gene_lengths_base = (gene_annot_v2['end'] - gene_annot_v2['start'])
# print(gene_lengths_base.head())

# # gene id (abbr) as index
gene_annot_file = PATH_GENEBODY_ANNOTATION
gene_annot = pd.read_csv(gene_annot_file, sep="\t")
gene_annot['gene_id_abbr'] = gene_annot['gene_id'].apply(lambda x: x.split('.')[0])
gene_annot = gene_annot.set_index('gene_id_abbr')

gene_lengths_base = (gene_annot['end'] - gene_annot['start'])
print(gene_lengths_base.head())

normalization_options = {
    'smarter_nuclei': 'TPM',
    'smarter_cells': 'TPM',
    'snatac_gene': 'TPM',
    '10x_nuclei_v3_macosko': 'CPM',
    '10x_cells_v3': 'CPM',
    '10x_nuclei_v3': 'CPM',
    'snmcseq_gene': 'MC',
    '10x_cells_v2': 'CPM',
#     '10x_nuclei_v2': 'CPM',
#     'smarter-cells-v1': 'TPM',
}

# highly variable features

for mod in mods_selected:
    ti = time.time()
    print(mod)
    
    # read metadata
    normalization_option = normalization_options[mod]
    f_meta = f_meta_format.format(SRC_DIR, mod) ##
    meta = pd.read_csv(f_meta, sep="\t", index_col=0)
    
    # read data matrix
    if normalization_option == 'MC':
        f_data_gene = f_data_format.format(SRC_DIR, mod, '', 'gene')
        f_data_cell = f_data_format.format(SRC_DIR, mod, '', 'cell')
        f_data_c = f_data_format.format(SRC_DIR, mod, 'CG_', 'npz')
        f_data_mc = f_data_format.format(SRC_DIR, mod, 'mCG_', 'npz')
        
        # read in files
        print(mod, "Reading in files {}".format(time.time()-ti))
        gxc_raw = snmcseq_utils.load_gc_matrix_methylation(f_data_gene, f_data_cell, f_data_mc, f_data_c)
        print(gxc_raw.data['mc'].shape, gxc_raw.data['c'].shape)
        print(time.time()-ti)
        
        # output file
        f_hvftr_data_methylation = f_hvftr_format.format(DST_DIR, mod, 'tsv') 
        print(time.time()-ti)
        
        # check meta cells agree with gxc cells
        assert np.all(meta.index.values == gxc_raw.cell)
        # check genes are uniq 
        assert len(gxc_raw.gene) == len(np.unique(gxc_raw.gene)) 
        
        # do
        gxc_hvftr = CEMBA_preproc_utils.preproc_methylation(
                                                            gxc_raw,
                                                            meta,
                                                            global_value_col=settings[mod].global_mean, 
                                                            base_call_cutoff=20, 
                                                            sufficient_coverage_fraction=0.95,
                                                            hv_percentile=30,
                                                            n_qcut=10,
                                                            )
        # save
        print(mod, "Saving to files {}".format(time.time()-ti))
        gxc_hvftr.to_csv(f_hvftr_data_methylation, sep="\t", header=True, index=True, na_rep='NA')
        
        
        
    else:
        f_data = f_data_format.format(SRC_DIR, mod, '', 'npz') 
        f_data_gene = f_data_format.format(SRC_DIR, mod, '', 'gene') 
        f_data_cell = f_data_format.format(SRC_DIR, mod, '', 'cell') 
        
        # read in files
        print(mod, "Reading in files {}".format(time.time()-ti))
        gxc_raw = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data) # checked dimensions in agreement internally
        print(gxc_raw.data.shape)
        
        # output files
        f_hvftr_data = f_hvftr_format.format(DST_DIR, mod, 'npz') 
        f_hvftr_gene = f_hvftr_format.format(DST_DIR, mod, 'gene') 
        f_hvftr_cell = f_hvftr_format.format(DST_DIR, mod, 'cell') 
        
        # check meta cells agree with gxc cells
        assert np.all(meta.index.values == gxc_raw.cell)
        # check genes are uniq 
        assert len(gxc_raw.gene) == len(np.unique(gxc_raw.gene)) 
    
        # get hvftrs
        print(mod, "Preproc and get highly variable genes {}".format(time.time()-ti))
        if normalization_option == 'CPM':
            gxc_hvftr = CEMBA_preproc_utils.preproc_rna_cpm_based(
                                             gxc_raw, 
                                             sufficient_cell_coverage=0.01, 
                                             hv_percentile=30, hv_ncut=10)
            # save
            print(mod, "Saving to files {}".format(time.time()-ti))
            snmcseq_utils.save_gc_matrix(gxc_hvftr, f_hvftr_gene, f_hvftr_cell, f_hvftr_data)

        elif normalization_option == 'TPM':
            gene_lengths = gene_lengths_base.reindex(gxc_raw.gene)
            gxc_hvftr = CEMBA_preproc_utils.preproc_rna_tpm_based(
                                             gxc_raw, gene_lengths, impute_gene_lengths=True, 
                                             sufficient_cell_coverage=0.01, 
                                             hv_percentile=30, hv_ncut=10)
            # save
            print(mod, "Saving to files {}".format(time.time()-ti))
            snmcseq_utils.save_gc_matrix(gxc_hvftr, f_hvftr_gene, f_hvftr_cell, f_hvftr_data)
    
    
    print(mod, "Total time used: {}".format(time.time()-ti))
    

## Check highly-variable genes

for mod in mods_selected:
    print(mod)
    if settings[mod].mod_category == 'mc':
        f_hvftr_data = f_hvftr_format.format(SRC_DIR, mod, 'tsv') 
        gxc_hvftr = pd.read_csv(f_hvftr_data, sep="\t", index_col=0)
        print(gxc_hvftr.index.values)
        print(gxc_hvftr.columns.values)
        print(gxc_hvftr.shape)
        has_nan = np.isnan(gxc_hvftr.values).any()
        print("Contains NaN? {}".format(has_nan))
        
        continue
        
    f_hvftr_data = f_hvftr_format.format(SRC_DIR, mod, 'npz') 
    f_hvftr_gene = f_hvftr_format.format(SRC_DIR, mod, 'gene') 
    f_hvftr_cell = f_hvftr_format.format(SRC_DIR, mod, 'cell') 
    gxc_hvftr = snmcseq_utils.load_gc_matrix(f_hvftr_gene, f_hvftr_cell, f_hvftr_data)
    print(gxc_hvftr.gene)
    print(gxc_hvftr.cell)
    print(len(gxc_hvftr.gene), len(gxc_hvftr.cell), gxc_hvftr.data.shape)
    has_nan = np.isnan(gxc_hvftr.data.data).any()
    print("Contains NaN? {}".format(has_nan))
#     break

### get and save ftrs 

for mod in mods_selected:
    ti = time.time()
    print(mod)
    
    # read metadata
    normalization_option = normalization_options[mod]
    f_meta = f_meta_format.format(SRC_DIR, mod) ##
    meta = pd.read_csv(f_meta, sep="\t", index_col=0)
    
    # read data matrix
    if normalization_option == 'MC':
        f_data_gene = f_data_format.format(SRC_DIR, mod, '', 'gene')
        f_data_cell = f_data_format.format(SRC_DIR, mod, '', 'cell')
        f_data_c = f_data_format.format(SRC_DIR, mod, 'CG_', 'npz')
        f_data_mc = f_data_format.format(SRC_DIR, mod, 'mCG_', 'npz')
        
        # read in files
        print(mod, "Reading in files {}".format(time.time()-ti))
        gxc_raw = snmcseq_utils.load_gc_matrix_methylation(f_data_gene, f_data_cell, f_data_mc, f_data_c)
        print(gxc_raw.data['mc'].shape, gxc_raw.data['c'].shape)
        print(time.time()-ti)
        
        # output file
        f_ftr_data_methylation = f_ftr_format.format(DST_DIR, mod, 'tsv') 
        print(time.time()-ti)
        
        # check meta cells agree with gxc cells
        assert np.all(meta.index.values == gxc_raw.cell)
        # check genes are uniq 
        assert len(gxc_raw.gene) == len(np.unique(gxc_raw.gene)) 
        
        ### do
        base_call_cutoff = 10
        sufficient_coverage_fraction = 0.80
        #  select genes covered (10 counts) in > 80% of cells
        n_gene, n_cell = gxc_raw.data['c'].shape
        gene_cov = (gxc_raw.data['c'] > base_call_cutoff).sum(axis=1)
        gene_cov = np.array(gene_cov).squeeze()/n_cell # fraction of cells covered
        cond = gene_cov>sufficient_coverage_fraction
        print(cond.sum())

        # to full matrix
        df_c = pd.DataFrame(
            gxc_raw.data['c'].tocsr()[cond, :].todense(),
            index=np.array(gxc_raw.gene)[cond],
            columns=gxc_raw.cell,
        )
        df_mc = pd.DataFrame(
            gxc_raw.data['mc'].tocsr()[cond, :].todense(),
            index=np.array(gxc_raw.gene)[cond],
            columns=gxc_raw.cell,
        )
        ### do
        
        # compute normalized methylation matrix 
        df_mcc = snmcseq_utils.get_mcc_lite_v2(df_c, df_mc, base_call_cutoff=base_call_cutoff)
        gxc_ftr = df_mcc.divide(meta.loc[df_mcc.columns.values, settings[mod].global_mean_mcg], axis=1)
        # save
        print(mod, "Saving to files {}".format(time.time()-ti))
        gxc_ftr.to_csv(f_ftr_data_methylation, sep="\t", header=True, index=True, na_rep='NA')
        
        
    else:
        f_data = f_data_format.format(SRC_DIR, mod, '', 'npz') 
        f_data_gene = f_data_format.format(SRC_DIR, mod, '', 'gene') 
        f_data_cell = f_data_format.format(SRC_DIR, mod, '', 'cell') 
        
        # read in files
        print(mod, "Reading in files {}".format(time.time()-ti))
        gxc_raw = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data) # checked dimensions in agreement internally
        print(gxc_raw.data.shape)
        
        # output files
        f_ftr_data = f_ftr_format.format(DST_DIR, mod, 'npz') 
        f_ftr_gene = f_ftr_format.format(DST_DIR, mod, 'gene') 
        f_ftr_cell = f_ftr_format.format(DST_DIR, mod, 'cell') 
        
        # check meta cells agree with gxc cells
        assert np.all(meta.index.values == gxc_raw.cell)
        # check genes are uniq 
        assert len(gxc_raw.gene) == len(np.unique(gxc_raw.gene)) 
    
        # get hvftrs
        print(mod, "Preproc and get highly variable genes {}".format(time.time()-ti))
        if normalization_option == 'CPM':
            # do 
            gxc_ftr = snmcseq_utils.sparse_logcpm(gxc_raw, mode='logcpm')
            # save
            print(mod, "Saving to files {}".format(time.time()-ti))
            snmcseq_utils.save_gc_matrix(gxc_ftr, f_ftr_gene, f_ftr_cell, f_ftr_data)

        elif normalization_option == 'TPM':
            # do
            gene_lengths = gene_lengths_base.reindex(gxc_raw.gene)
            gene_lengths = gene_lengths.fillna(np.nanmean(gene_lengths.values))
            gxc_ftr = snmcseq_utils.sparse_logtpm(gxc_raw, gene_lengths)
            # save
            print(mod, "Saving to files {}".format(time.time()-ti))
            snmcseq_utils.save_gc_matrix(gxc_ftr, f_ftr_gene, f_ftr_cell, f_ftr_data)
    
    
    print(mod, "Total time used: {}".format(time.time()-ti))
    

print(gxc_ftr.shape)
gxc_ftr.head()

f = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_neurons/snmcseq_gene_features_mcg.tsv'
gxc_ftr.to_csv(f, sep="\t", header=True, index=True, na_rep='NA')
