==> fig4/02.prep_data_qc_cells.ipynb <==
# Organize freezed data
Aug 21th, 2019
Fangming Xie

- use reformatted data as input
- extract QC'ed cells only
- make metadata and data consistent
- unique cell name for each dataset: dataset + '_' + cellname (10x_nuclei_v3)

import sys
sys.path.insert(0, '/cndd/fangming/CEMBA/snmcseq_dev')
import importlib

from __init__ import *
from __init__jupyterlab import *
import snmcseq_utils
importlib.reload(snmcseq_utils)
import CEMBA_clst_utils
from scipy.io import mmread
from scipy import sparse
import re
import time


def abbr_cell_names(long_name):
    """Turn a long name: "$barcode-$number$substring" into "$barcode-$number"
    """
    barcode, suffix = long_name.split('-')
    match = re.match(r'^\d+', suffix)
    suffix_short = suffix[match.start():match.end()]
    short_name = '{}-{}'.format(barcode, suffix_short)
    return short_name

SRC_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_formatted'
DST_DIR = '/cndd/fangming/CEMBA/data/MOp_all/data_freeze_qc_cells'

f = PATH_GENEBODY_ANNOTATION 
df_genes = pd.read_csv(f, sep="\t")
df_genes['gene_id'] = df_genes['gene_id'].apply(lambda x: x.split('.')[0])
df_genes = df_genes.set_index('gene_id')
df_genes_v2 = df_genes.reset_index().groupby('gene_name').first()
print(df_genes.shape)
df_genes.head()
print(df_genes_v2.shape)
df_genes_v2.head()

### 10x_nuclei_v3_macosko 
- 2 metadata .csv
- 1 data matrix in .rds format
- concatenate male and female, get autosomo genes only

dataset_name = '10x_nuclei_v3_macosko'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data = os.path.join(SRC_DIR, dataset_name+'_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data = os.path.join(DST_DIR, dataset_name+'_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(fout_meta)
print(f_data)
print(fout_data)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.shape)
meta.head()

gc_mat = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data)
print(gc_mat.data.shape, gc_mat.gene.shape, gc_mat.cell.shape)

# select cells

selected_cells = meta.index.values
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells)
selected_cells = dataset_name + '_' + selected_cells

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, gc_mat.data.tocsc()[:, selected_cells_id])


# output
# save 
ti = time.time()
snmcseq_utils.save_gc_matrix(gc_mat_update, fout_data_gene, fout_data_cell, fout_data)
print(time.time()-ti)

# assemble metadata (keep order)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)

### 10x_nuclei_v3

- gene_ids
- all cells passed QC (QC.csv) metadata

- all cells data
- output

dataset_name = '10x_nuclei_v3'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data = os.path.join(SRC_DIR, dataset_name+'_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data = os.path.join(DST_DIR, dataset_name+'_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(fout_meta)
print(f_data)
print(fout_data)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.shape)
meta.head()

gc_mat = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data)
print(gc_mat.data.shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([abbr_cell_names(cell) for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells

# check if the reduced names are unique
assert len(np.unique(selected_cells) == len(np.unique(selected_cells_abbr)))
# check if new names mapped to short ones
assert np.all(selected_cells_id != -1)

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, gc_mat.data.tocsc()[:, selected_cells_id])
print(gc_mat_update.data.shape, gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# output
# save 
ti = time.time()
snmcseq_utils.save_gc_matrix(gc_mat_update, fout_data_gene, fout_data_cell, fout_data)
print(time.time()-ti)

# assemble metadata (keep order, add male and female)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)

### 10x_cells_v3 

dataset_name = '10x_cells_v3'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data = os.path.join(SRC_DIR, dataset_name+'_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data = os.path.join(DST_DIR, dataset_name+'_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(fout_meta)
print(f_data)
print(fout_data)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.shape)
meta.head()

gc_mat = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data)
print(gc_mat.data.shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([abbr_cell_names(cell) for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells

# check if the reduced names are unique
assert len(np.unique(selected_cells) == len(np.unique(selected_cells_abbr)))
# check if new names mapped to short ones
assert np.all(selected_cells_id != -1)

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, gc_mat.data.tocsc()[:, selected_cells_id])
print(gc_mat_update.data.shape, gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# output
# save 
ti = time.time()
snmcseq_utils.save_gc_matrix(gc_mat_update, fout_data_gene, fout_data_cell, fout_data)
print(time.time()-ti)

# assemble metadata (keep order, add male and female)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)

### 10x_cells_v2 

dataset_name = '10x_cells_v2'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data = os.path.join(SRC_DIR, dataset_name+'_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data = os.path.join(DST_DIR, dataset_name+'_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(fout_meta)
print(f_data)
print(fout_data)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.shape)
meta.head()

a = np.sort(np.unique([cell.split('-')[1] 
              for cell in meta.index.values]))

b = np.sort(np.unique([cell.split('-')[1]
           for cell in gc_mat.cell]))
print(a, b)

gc_mat = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data)
print(gc_mat.data.shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([abbr_cell_names(cell) for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells

# check if the reduced names are unique
assert len(np.unique(selected_cells) == len(np.unique(selected_cells_abbr)))
# check if new names mapped to short ones
assert np.all(selected_cells_id != -1)

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, gc_mat.data.tocsc()[:, selected_cells_id])
print(gc_mat_update.data.shape, gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# output
# save 
ti = time.time()
snmcseq_utils.save_gc_matrix(gc_mat_update, fout_data_gene, fout_data_cell, fout_data)
print(time.time()-ti)

# assemble metadata (keep order, add male and female)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)

### Smart-cells

dataset_name = 'smarter_cells'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data = os.path.join(SRC_DIR, dataset_name+'_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data = os.path.join(DST_DIR, dataset_name+'_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(fout_meta)
print(f_data)
print(fout_data)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.shape)
meta.head()

gc_mat = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data)
print(gc_mat.data.shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([cell for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells
assert np.all(selected_cells_id != -1)

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, gc_mat.data.tocsc()[:, selected_cells_id])
print(gc_mat_update.data.shape, gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# output
# save 
ti = time.time()
snmcseq_utils.save_gc_matrix(gc_mat_update, fout_data_gene, fout_data_cell, fout_data)
print(time.time()-ti)

# assemble metadata (keep order, add male and female)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)

### Smarter-nuclei

dataset_name = 'smarter_nuclei'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data = os.path.join(SRC_DIR, dataset_name+'_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data = os.path.join(DST_DIR, dataset_name+'_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(fout_meta)
print(f_data)
print(fout_data)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.shape)
meta.head()

gc_mat = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data)
print(gc_mat.data.shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([cell for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells
assert np.all(selected_cells_id != -1)

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, gc_mat.data.tocsc()[:, selected_cells_id])
print(gc_mat_update.data.shape, gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# output
# save 
ti = time.time()
snmcseq_utils.save_gc_matrix(gc_mat_update, fout_data_gene, fout_data_cell, fout_data)
print(time.time()-ti)

# assemble metadata (keep order, add male and female)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)

### mC

- QC'ed cells
- remove version on ensembl id

dataset_name = 'snmcseq_gene'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data_mc = os.path.join(SRC_DIR, dataset_name+'_mCH_raw.npz') 
f_data_c = os.path.join(SRC_DIR, dataset_name+'_CH_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(f_data_mc)
print(f_data_c)
print(f_data_gene)
print(f_data_cell)

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data_mc = os.path.join(DST_DIR, dataset_name+'_mCH_raw.npz') 
fout_data_c = os.path.join(DST_DIR, dataset_name+'_CH_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 


print(fout_meta)
print(fout_data_mc)
print(fout_data_c)
print(fout_data_gene)
print(fout_data_cell)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
# # meta = meta[meta['MajorCluster']!='Outlier']
# print(meta.groupby('MajorCluster').size())
print(meta.shape)
meta.head()

gc_mat = snmcseq_utils.load_gc_matrix_methylation(f_data_gene, f_data_cell, f_data_mc, f_data_c)
print(gc_mat.data['mc'].shape, gc_mat.data['c'].shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([cell for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells
assert np.all(selected_cells_id != -1)

data_update = {'mc': gc_mat.data['mc'].tocsc()[:, selected_cells_id], 
               'c': gc_mat.data['c'].tocsc()[:, selected_cells_id], 
              }

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, data_update)
print(gc_mat_update.data['mc'].shape, 
      gc_mat_update.data['c'].shape, 
      gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# save mc and c tables

snmcseq_utils.save_gc_matrix_methylation(gc_mat_update, 
                                         fout_data_gene, fout_data_cell, 
                                         fout_data_mc, fout_data_c)

# assemble metadata (keep order, add male and female)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)

### ATAC 
- rename barcode and header
- read snap files

dataset_name = 'snatac_gene'

f_meta = os.path.join(SRC_DIR, dataset_name+'_metadata.tsv') 
f_data = os.path.join(SRC_DIR, dataset_name+'_raw.npz') 
f_data_gene = os.path.join(SRC_DIR, dataset_name+'_raw.gene') 
f_data_cell = os.path.join(SRC_DIR, dataset_name+'_raw.cell') 

fout_meta = os.path.join(DST_DIR, dataset_name+'_metadata.tsv') 
fout_data = os.path.join(DST_DIR, dataset_name+'_raw.npz') 
fout_data_gene = os.path.join(DST_DIR, dataset_name+'_raw.gene') 
fout_data_cell = os.path.join(DST_DIR, dataset_name+'_raw.cell') 

print(f_meta)
print(fout_meta)
print(f_data)
print(fout_data)

meta = pd.read_csv(f_meta, sep="\t", index_col=0)
print(meta.groupby('cluster').size())
print(meta.shape)
meta.head()

gc_mat = snmcseq_utils.load_gc_matrix(f_data_gene, f_data_cell, f_data)
print(gc_mat.data.shape, gc_mat.gene.shape, gc_mat.cell.shape)

selected_cells = meta.index.values
selected_cells_abbr = np.array([cell for cell in selected_cells])
selected_cells_id = snmcseq_utils.get_index_from_array(gc_mat.cell, selected_cells_abbr)
selected_cells = dataset_name + '_' + selected_cells
assert np.all(selected_cells_id != -1)

gc_mat_update = GC_matrix(gc_mat.gene, selected_cells, gc_mat.data.tocsc()[:, selected_cells_id])
print(gc_mat_update.data.shape, gc_mat_update.gene.shape, gc_mat_update.cell.shape)

# output
# save 
ti = time.time()
snmcseq_utils.save_gc_matrix(gc_mat_update, fout_data_gene, fout_data_cell, fout_data)
print(time.time()-ti)

# assemble metadata (keep order, add male and female)
meta.index = selected_cells
meta.index.name = 'cell'
meta.to_csv(fout_meta, sep="\t", header=True, index=True)
